+++ CHANGES.md
@@ -13,6 +13,7 @@
   your preferred path and then rebuild all assets (e.g. `make dist`).
 - Use `listenTo` to avoid memory leaks when views get removed.
 - #1692 Bugfix: `TypeError: oldest_message is undefined`
+- #1704 SECURITY FIX: Impersonation by misusage of groupchat carbons
 - #1705 Bugfix: `this.roomspanel` is `undefined` after hibernating